src/requirements.txt
Django==1.10.4
djangorestframework==3.5.3
httpie==0.9.6
pkg-resources==0.0.0
Pygments==2.1.3
requests==2.12.3
